refactor: CustomModal作成
