+++ React/react_todo_list/src/components/DeleteConfirmModal.tsx
@@ -1,45 +1,23 @@
 import { FC } from "react";
-import Modal from 'react-modal';
+// import Modal from 'react-modal';
 import { ModalType } from "../types/modal" 
+import { CustomModal } from "./organisms/CustomModal" 
 
 type DeleteConfirmModalType = {
   modal: ModalType;
-  closeModal: () => void;
   onClickDeleteTodo: () => void;
+  closeModal: () => void;
 }
 
-export const DeleteConfirmModal: FC<DeleteConfirmModalType> = ({ modal, closeModal, onClickDeleteTodo }) => {
-  const customStyles = {
-    content: {
-      top: '25%',
-      left: '50%',
-      right: 'auto',
-      bottom: 'auto',
-      width: '500px',
-      height: '120px',
-      marginRight: '-50%',
-      transform: 'translate(-50%, -50%)',
-    },
-  }
+export const DeleteConfirmModal: FC<DeleteConfirmModalType> = ({ modal, onClickDeleteTodo, closeModal }) => {
+  const confirmText = ""
 
   return (
-    <Modal
-      isOpen={modal.isOpen}
-      onRequestClose={closeModal}
-      style={customStyles}
-      contentLabel="Delete Modal"
-    >
-      <p className="mb-4">本当に削除してもよろしいですか？</p>
-      <div className="flex justify-end">
-        <button
-          className="w-24 p-2 text-sm border border-blue-500 rounded-md bg-blue-500 text-white"
-          onClick={onClickDeleteTodo}
-        >OK</button>
-        <button
-          className="w-24 p-2 ml-2 text-sm border border-gray-200 rounded-md text-blue-500"
-          onClick={closeModal}
-        >キャンセル</button>
-      </div>
-    </Modal>
+    <CustomModal
+      modal={modal}
+      confirmText={confirmText}
+      confirm={onClickDeleteTodo}
+      cancel={closeModal}
+    />
   )
 }
